Add the ability to log to a file
Add a dependency on tracing-appender

For reasons I don't know, tracing-appender didn't support custom suffix.
This means it's impossible to log to a file with a .log extension. I
submitted a fix that was merged to add support for it, but there has
been no release of tracing-appender in almost a year

Closes: #5233